% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines.R
\name{age_group_eligible_for_dose_vaccine_new}
\alias{age_group_eligible_for_dose_vaccine_new}
\title{Find individuals in each age group eligible for a dose (multi-dose, no types)}
\usage{
age_group_eligible_for_dose_vaccine_new(dose, parameters, variables, t, dt)
}
\arguments{
\item{dose}{which dose?}

\item{parameters}{a list}

\item{variables}{a list}

\item{t}{the current time step}

\item{dt}{size of the time step}
}
\value{
a list containing a \code{\link[individual]{Bitset}} for each age bin
}
\description{
Find those individuals who have had the dose preceding \code{dose},
have not yet received the next one, and are beyond the \code{dose_period}.
This is similar to the function \code{\link[nimue]{eligable_for_second}} in the nimue package.
This function should only be called if simulation time is greater than the \code{dose_period}
}
