% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_hospital.R
\name{allocate_treatment}
\alias{allocate_treatment}
\title{Allocate treatment}
\usage{
allocate_treatment(variables, need_treatment, treated_state, limit)
}
\arguments{
\item{variables}{Model variables}

\item{need_treatment}{a vector of individuals who need treatment}

\item{treated_state}{a list states for individuals receiving treatment}

\item{limit}{the number of individuals who can receive treatment}
}
\description{
sample a subset of individuals who will receive treatment. The subset is
allways smaller than the limit of treatments available minus those already
receiving treatment
}
