% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vaccination.R
\name{assign_doses}
\alias{assign_doses}
\title{Assign N doses to age groups based on weightings by number of people eligible to be vaccinated (multi-dose, no types)}
\usage{
assign_doses(t, dt, doses, n_to_cover, variables, events, phase, parameters)
}
\arguments{
\item{doses}{Total available doses}

\item{n_to_cover}{Number of people eligible to be vaccinated in each age group, from \code{\link{target_pop}}}
}
\description{
Please make sure you are subtracting these doses from some daily total outside of this function.
It combines functionality of \code{\link[nimue]{assign_doses}} and \code{\link[nimue]{administer_first_dose}}/\code{\link[nimue]{administer_second_dose}}.
}
