% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{attach_event_listeners}
\alias{attach_event_listeners}
\title{Attach listeners to events}
\usage{
attach_event_listeners(variables, events, parameters, dt)
}
\arguments{
\item{variables}{list of variables in the model}

\item{events}{a list of events in the model}

\item{parameters}{the model parameters}

\item{dt}{size of time step}
}
\description{
defines processes for events that can be scheduled in the future
}
