% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_nat.R
\name{attach_event_listeners_independent_nat}
\alias{attach_event_listeners_independent_nat}
\title{Attach event listeners for modeling independent infection-derived NAT}
\usage{
attach_event_listeners_independent_nat(variables, events, parameters, dt)
}
\arguments{
\item{variables}{a named list of variables}

\item{events}{a named list of events}

\item{parameters}{the parameters}

\item{dt}{size of time step}
}
\description{
Attach event listeners for modeling independent infection-derived NAT
}
