% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_vaccination_nimue.R
\name{attach_event_listeners_nimue}
\alias{attach_event_listeners_nimue}
\title{Attach listeners to events (nimue vaccine model)}
\usage{
attach_event_listeners_nimue(variables, events, parameters, dt, shift = 1L)
}
\arguments{
\item{variables}{list of variables in the model}

\item{events}{a list of events in the model}

\item{parameters}{the model parameters}

\item{dt}{size of time step}

\item{shift}{schedule future events after minimum number of time step delay}

\item{shift_exposure}{schedule exposure event after minimum number of time step delay}
}
\description{
defines processes for events that can be scheduled in the future
}
