% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_vaccination.R
\name{attach_event_listeners_vaccination}
\alias{attach_event_listeners_vaccination}
\title{Attach event listeners for vaccination events (multi-dose, no types)}
\usage{
attach_event_listeners_vaccination(variables, events, parameters, dt)
}
\arguments{
\item{variables}{a named list of variables}

\item{events}{a named list of events}

\item{parameters}{the parameters}

\item{type}{character type of vaccine}
}
\description{
Attach event listeners for vaccination events (multi-dose, no types)
}
