% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-hosp.R
\name{attach_hosp_listeners}
\alias{attach_hosp_listeners}
\title{Attach listeners to track hospitalization and ICU incidence}
\usage{
attach_hosp_listeners(renderers, events)
}
\arguments{
\item{renderers}{a named list of \link[individual:Render]{individual::Render} objects}

\item{events}{a named list of \link[individual:TargetedEvent]{individual::TargetedEvent} objects}
}
\description{
Attach listeners to track hospitalization and ICU incidence
}
