% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{attach_tracking_listener_age_incidence}
\alias{attach_tracking_listener_age_incidence}
\title{Attach listener to track age-stratified incidence}
\usage{
attach_tracking_listener_age_incidence(events, renderer, age, parameters)
}
\arguments{
\item{events}{a list of events}

\item{renderer}{a \link[individual:Render]{individual::Render} object}

\item{age}{an \link[individual:IntegerVariable]{individual::IntegerVariable} object}

\item{parameters}{model parameters (\link{get_parameters})}
}
\description{
To use age-stratified incidence tracking, run the lines in examples
anytime after all events have been made and initialized but before
simulation begins.
}
\examples{
\dontrun{
incidence_renderer <- individual::Render$new(timesteps)
attach_tracking_listener_age_incidence(
    events = events, renderer = age_incidence_renderer,
    age = variables$discrete_age, parameters = parameters
 )
}
}
