% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{attach_tracking_listener_incidence}
\alias{attach_tracking_listener_incidence}
\title{Attach listener to track incidence}
\usage{
attach_tracking_listener_incidence(events, renderer)
}
\arguments{
\item{events}{a list of events}

\item{renderer}{a \code{\link[individual]{Render}} object}
}
\description{
To use incidence tracking, run the lines in examples
anytime after all events have been made and initialized and before
simulation begins.
}
\examples{
\dontrun{
incidence_renderer <- individual::Render$new(timesteps)
attach_tracking_listener_incidence(events = events,renderer = incidence_renderer)
}
}
