% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{categorical_count_renderer_process_daily}
\alias{categorical_count_renderer_process_daily}
\title{Render categories every day}
\usage{
categorical_count_renderer_process_daily(renderer, variable, categories, dt)
}
\arguments{
\item{renderer}{a \code{\link[individual]{Render}} object}

\item{variable}{a \code{\link[individual]{CategoricalVariable}} object}

\item{categories}{a character vector of categories to render}

\item{dt}{size of time step}
}
\value{
a function which can be passed as a process to \code{\link{simulation_loop}}
}
\description{
Renders the number of individuals in each category, but only
on timesteps that correspond to a day. This saves memory and potentially time
for simulations with small timesteps.
}
\details{
#' @title Render double variables every day
#' @description This only renders output on timesteps
#' that correspond to a day. This saves memory and potentially time for
#' simulations with small timesteps.
#' @param renderer a \code{\link[individual]{Render}} object
#' @param variable a \code{\link[individual]{DoubleVariable}} object
#' @param dt size of time step
#' @export
double_count_render_process_daily <- function(renderer, variable, dt) {
stopifnot(inherits(variable, "DoubleVariable"))
stopifnot(inherits(renderer, "Render"))
function(t) {
if ((t * dt) \%\% 1 == 0) {
values <- variable$get_values()
renderer$render(name = as.character(1:length(values)), value = values,timestep = as.integer(t * dt))
}
}
}
}
