% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{create_age_incidence_tracking_listener}
\alias{create_age_incidence_tracking_listener}
\title{Create listener to track age-stratified incidence}
\usage{
create_age_incidence_tracking_listener(renderer, age, parameters)
}
\arguments{
\item{renderer}{a \link[individual:Render]{individual::Render} object}

\item{age}{an \link[individual:IntegerVariable]{individual::IntegerVariable} object}

\item{parameters}{model parameters (\link{get_parameters})}
}
\description{
Create listener to track age-stratified incidence
}
