% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{create_age_variables}
\alias{create_age_variables}
\title{Create age variables}
\usage{
create_age_variables(pop, parameters, continuous = FALSE)
}
\arguments{
\item{pop}{population list}

\item{parameters}{model parameters}

\item{continuous}{return both age by year and bin?}
}
\value{
named list of variable objects
}
\description{
Create individual variables for continuous and discrete age
}
