% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{create_continuous_age_variable}
\alias{create_continuous_age_variable}
\title{Continuous age variable}
\usage{
create_continuous_age_variable(pop, max_age = 100)
}
\arguments{
\item{pop}{population list}

\item{max_age}{maximum age to be drawn}
}
\value{
continuous age variable
}
\description{
Create a continuous age variable for the population
}
