% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{create_discrete_age_variable}
\alias{create_discrete_age_variable}
\title{Discrete age variable}
\usage{
create_discrete_age_variable(ages, pop)
}
\arguments{
\item{ages}{Vector of ages from \link{create_continuous_age_variable}}

\item{pop}{Vector of integer ages}
}
\value{
discrete age variable
}
\description{
Create a discrete age variable for each of the
length(pop$age_group) distinct age groups
}
