% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_event_scheduler_listener}
\alias{create_event_scheduler_listener}
\title{A listener to schedule future events}
\usage{
create_event_scheduler_listener(event, duration, func, shift, dt)
}
\arguments{
\item{event}{the future event to be schedule}

\item{duration}{mean duration of waiting time to be scheduled}

\item{func}{either \code{\link{make_rerlang}} or \code{\link{make_rexp}}}

\item{shift}{add integer number of time steps to sampled value}

\item{dt}{size of time step}
}
\description{
a listener to be attached to a \code{\link[individual]{TargetedEvent}}
to schedule a future event when that event fires.
}
