% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_events}
\alias{create_events}
\title{Create events}
\usage{
create_events(parameters, vaccines = NULL)
}
\arguments{
\item{parameters}{a named list of individual::Event}

\item{vaccines}{not currently used}
}
\description{
Create events
}
