% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_vaccination_nimue.R
\name{create_events_nimue}
\alias{create_events_nimue}
\title{Append vaccination events (nimue vaccine model)}
\usage{
create_events_nimue(events, parameters)
}
\arguments{
\item{events}{a named list of individual::Event}

\item{parameters}{write me!}
}
\description{
Append vaccination events (nimue vaccine model)
}
