% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_vaccination.R
\name{create_events_vaccination}
\alias{create_events_vaccination}
\title{Append vaccination events}
\usage{
create_events_vaccination(events, parameters)
}
\arguments{
\item{events}{a named list of individual::Event}

\item{parameters}{write me!}
}
\description{
Append vaccination events
}
