% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_exposure_variants.R
\name{create_exposure_scheduler_listener_vaccine_voc}
\alias{create_exposure_scheduler_listener_vaccine_voc}
\title{Modelling the progression to either IMild, ICase, IAsymp with variants of concern (vaccine model, multi-dose, no types)}
\usage{
create_exposure_scheduler_listener_vaccine_voc(
  events,
  variables,
  parameters,
  dt,
  shift = 0
)
}
\arguments{
\item{events}{a list of events in the model}

\item{variables}{the available human variables}

\item{parameters}{model parameters from \code{\link{get_parameters_nimue}}}

\item{dt}{the time step}

\item{shift}{passed to \code{\link{make_rerlang}}}
}
\description{
Age dependent outcome of exposure
}
