% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-hosp.R
\name{create_hosp_renderers}
\alias{create_hosp_renderers}
\title{Create list of renderer objects to track hospitalization and ICU incidence}
\usage{
create_hosp_renderers(parameters)
}
\arguments{
\item{parameters}{named list of model parameters}
}
\value{
a named list of \link[individual:Render]{individual::Render} objects
}
\description{
Create list of renderer objects to track hospitalization and ICU incidence
}
