% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_hospital.R
\name{create_hospital_scheduler_listener_cpp}
\alias{create_hospital_scheduler_listener_cpp}
\title{Create C++ listener function to schedule events upon hospitilisation}
\usage{
create_hospital_scheduler_listener_cpp(parameters, variables, events)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{a list of all of the model variables}

\item{events}{a list of all of the model events}
}
\description{
When the \code{hospitilisation} event fires, this listener should be called
to schedule future events and state changes for those persons. This calls
the function \link{create_hospital_scheduler_listener_cpp_internal} to make
an \linkS4class{externalptr} object.
}
