% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_hospital_scheduler_listener_cpp_internal}
\alias{create_hospital_scheduler_listener_cpp_internal}
\title{Internal C++ function factory for scheduling events upon hospitilisation}
\usage{
create_hospital_scheduler_listener_cpp_internal(
  parameters,
  states,
  discrete_age,
  imv_get_die,
  imv_get_live,
  imv_not_get_die,
  imv_not_get_live,
  iox_get_die,
  iox_get_live,
  iox_not_get_die,
  iox_not_get_live
)
}
\arguments{
\item{parameters}{a named \link{list}}

\item{states}{a \link[individual:CategoricalVariable]{individual::CategoricalVariable} object}

\item{discrete_age}{a \link[individual:IntegerVariable]{individual::IntegerVariable} object}

\item{imv_get_die}{a \link[individual:TargetedEvent]{individual::TargetedEvent} object}

\item{imv_get_live}{a \link[individual:TargetedEvent]{individual::TargetedEvent} object}

\item{imv_not_get_die}{a \link[individual:TargetedEvent]{individual::TargetedEvent} object}

\item{imv_not_get_live}{a \link[individual:TargetedEvent]{individual::TargetedEvent} object}

\item{iox_get_die}{a \link[individual:TargetedEvent]{individual::TargetedEvent} object}

\item{iox_get_live}{a \link[individual:TargetedEvent]{individual::TargetedEvent} object}

\item{iox_not_get_die}{a \link[individual:TargetedEvent]{individual::TargetedEvent} object}

\item{iox_not_get_live}{a \link[individual:TargetedEvent]{individual::TargetedEvent} object}
}
\description{
Internal C++ function factory for scheduling events upon hospitilisation
}
