% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{create_incidence_tracking_listener}
\alias{create_incidence_tracking_listener}
\title{Create listener to track incidence}
\usage{
create_incidence_tracking_listener(renderer)
}
\arguments{
\item{renderer}{a \code{\link[individual]{Render}} object}
}
\description{
Create listener to track incidence
}
