% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_nat.R
\name{create_independent_nat_variables}
\alias{create_independent_nat_variables}
\title{Create independent infection-derived NAT variables}
\usage{
create_independent_nat_variables(variables, parameters)
}
\arguments{
\item{variables}{a list}

\item{parameters}{list of model parameters}
}
\value{
named list of individual::Variable
}
\description{
This creates an independent set of variables to track infection-derived NAT independently
from vaccine-derived NAT.
}
