% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_naturalimmunity.R
\name{create_natural_immunity_variables}
\alias{create_natural_immunity_variables}
\title{Create natural immunity variables}
\usage{
create_natural_immunity_variables(variables, parameters)
}
\arguments{
\item{variables}{a list}

\item{parameters}{list of model parameters}
}
\value{
named list of individual::Variable
}
\description{
Create natural immunity variables
}
