% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_state_update_listener}
\alias{create_state_update_listener}
\title{A listener to update state}
\usage{
create_state_update_listener(states, destination)
}
\arguments{
\item{states}{a \code{\link[individual]{CategoricalVariable}} object}

\item{destination}{the destination state}
}
\description{
a listener to be attached to a \code{\link[individual]{TargetedEvent}}
to update state when that event fires.
}
