% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_states.R
\name{create_state_variables}
\alias{create_state_variables}
\title{Define model states}
\usage{
create_state_variables(psq)
}
\arguments{
\item{psq}{the model parameters}
}
\value{
list of states
}
\description{
Create_states creates and initialises the human states for the model
}
