% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_vaccination_nimue.R
\name{create_v1v2_to_v3v4_listener_nimue}
\alias{create_v1v2_to_v3v4_listener_nimue}
\title{Create listener for start of vaccine protection (nimue vaccine model)}
\usage{
create_v1v2_to_v3v4_listener_nimue(
  variables,
  events,
  parameters,
  func,
  shift,
  dt
)
}
\arguments{
\item{variables}{a named list of variables}

\item{events}{a named list of individual::Event}

\item{parameters}{the parameters}

\item{func}{function to draw waiting times from}

\item{shift}{number of time steps to shift scheduled event by}

\item{dt}{size of time step}
}
\description{
This is called when entering v3v4 state
}
