% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_vaccination.R
\name{create_vaccine_variables}
\alias{create_vaccine_variables}
\title{Create vaccination variables (multi-dose, no types)}
\usage{
create_vaccine_variables(variables, pop, max_dose = 2)
}
\arguments{
\item{variables}{a list}

\item{pop}{population list}

\item{max_dose}{maximum number of possible doses}
}
\value{
named list of individual::Variable
}
\description{
Create all individual variables for humans
}
