% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_variants.R
\name{create_voc_variables}
\alias{create_voc_variables}
\title{Create variant of concern variables}
\usage{
create_voc_variables(variables, parameters)
}
\arguments{
\item{variables}{a list}

\item{parameters}{list of model parameters}
}
\value{
named list of individual::Variable
}
\description{
Please note the variant model assumes the user is applying
the vaccine model with antibody dynamics (not nimue or squire versions).
}
