% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cross_tab_compartments_age}
\alias{cross_tab_compartments_age}
\title{Cross tabulate compartments and age}
\usage{
cross_tab_compartments_age(compartments, age, num_ages, compartment_names)
}
\arguments{
\item{age}{a \link[individual:IntegerVariable]{individual::IntegerVariable}}

\item{num_ages}{number of age groups}

\item{doses}{a \link[individual:CategoricalVariable]{individual::CategoricalVariable}}

\item{num_doses}{number of doses}
}
\description{
Cross tabulate compartments and age
}
\examples{
\dontrun{
a <- IntegerVariable$new(0:4)
b <- IntegerVariable$new(c(1,2,3,1,2))
cross_tab_doses_age(a$.variable,b$.variable,4,3)
table(a$get_values(), b$get_values())
}
}
