% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cross_tab_doses_age}
\alias{cross_tab_doses_age}
\title{Cross tabulate doses and age}
\usage{
cross_tab_doses_age(doses, age, num_doses, num_ages)
}
\arguments{
\item{doses}{a \code{\link[individual]{IntegerVariable}}}

\item{age}{a \code{\link[individual]{IntegerVariable}}}

\item{num_doses}{number of doses}

\item{num_ages}{number of age groups}
}
\description{
The input vectors \code{doses} and \code{age}
must have the same number of values, this function does no argument checking.
}
\examples{
a <- IntegerVariable$new(0:4)
b <- IntegerVariable$new(c(1,2,3,1,2))
cross_tab_doses_age(a$.variable,b$.variable,4,3)
table(a$get_values(), b$get_values())
}
