% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{swap_ages}
\alias{swap_ages}
\title{Swap values that have been identified}
\usage{
swap_ages(swaps, age)
}
\arguments{
\item{swaps}{values to be swapped}

\item{age}{discrete age or continuous age}
}
\value{
age
}
\description{
Swap values that have been identified
}
