% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_asymptomatic}
\alias{get_asymptomatic}
\title{Function to add asymptomatic information}
\usage{
get_asymptomatic()
}
\value{
list of asymptomatic parameters
}
\description{
Function to add asymptomatic information
}
