% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_beta_cpp}
\alias{get_beta_cpp}
\title{Get a beta value}
\usage{
get_beta_cpp(beta_set, i)
}
\arguments{
\item{beta_set}{the set of beta values}

\item{i}{the day (assumes zero indexing)}
}
\description{
Get a beta value at some specific day
}
