% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_contact_matrix_cpp}
\alias{get_contact_matrix_cpp}
\title{Get contact matrix}
\usage{
get_contact_matrix_cpp(array, i)
}
\arguments{
\item{array}{the mixing matrix array (days x age x age)}

\item{i}{the day (indexes the first dimension, assumes zero indexing)}
}
\description{
Get the contact matrix at some specific day (1st dimension of array).
}
