% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines_fix.R
\name{get_current_coverage}
\alias{get_current_coverage}
\title{Get people who either have this dose or are scheduled for it}
\usage{
get_current_coverage(variables, events, dose, parameters)
}
\arguments{
\item{variables}{a list of model variables}

\item{events}{a list of model events}

\item{dose}{what dose to check coverage for}

\item{parameters}{a list of parameters}
}
\description{
Return a list of bitsets, one for each age group, containing.
those persons who either have had this dose or are scheduled for it.
The complement of each of these sets will be those
people who haven't gotten this dose and aren't scheduled for it.
}
