% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines.R
\name{get_current_eligible_from_coverage}
\alias{get_current_eligible_from_coverage}
\title{Get persons who are eligible for this dose}
\usage{
get_current_eligible_from_coverage(
  timestep,
  dt,
  coverage,
  variables,
  dose,
  parameters
)
}
\arguments{
\item{timestep}{the current time step}

\item{dt}{time step size}

\item{coverage}{output of \code{\link{get_current_coverage}}}

\item{variables}{a list of model variables}

\item{dose}{what dose to get eligibility for}

\item{parameters}{a list of parameters}
}
\description{
Get people who are eligible for this dose, from the output of \code{\link{get_current_coverage}}.
Because \code{coverage} is people (by age) who either have this dose or are scheduled for it,
if we are checking eligibility for dose 1 we just want the complement of those people. For doses > 1 we need to
first take the complement and then check
to see who of those persons got their previous dose past the waiting threshold.
This function returns a list of bitsets.
}
