% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vaccination.R
\name{get_current_prioritization_step}
\alias{get_current_prioritization_step}
\title{Get current prioritization step for a specific dose}
\usage{
get_current_prioritization_step(variables, parameters, dose)
}
\arguments{
\item{variables}{a list}

\item{parameters}{Model parameters}

\item{dose}{which dose?}
}
\description{
Get current prioritization step for a specific dose
}
