% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from squire model}
\usage{
get_parameters(
  iso3c = NULL,
  population = NULL,
  contact_matrix_set = NULL,
  time_period = 365,
  max_age = 100,
  dt,
  ...
)
}
\arguments{
\item{iso3c}{Character for country iso3c}

\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{contact_matrix_set}{Contact matrices used in simulation. Default =
NULL, which will generate this based on the \code{country}.}

\item{time_period}{Length of simulation. Default = 365}

\item{max_age}{the maximum age for humans}

\item{dt}{Time Step. Default = 0.1}

\item{...}{Other parameters for \link[squire:parameters_explicit_SEEIR]{squire::parameters_explicit_SEEIR}}
}
\value{
squire model parameters
}
\description{
Get parameters from squire model
}
