% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_nimue.R
\name{get_parameters_nimue}
\alias{get_parameters_nimue}
\title{Get parameters for safir (nimue vaccine model)}
\usage{
get_parameters_nimue(
  population,
  contact_mat,
  time_period,
  max_age = 100,
  lambda_external = NULL,
  dt = 1,
  ...
)
}
\arguments{
\item{population}{integer vector of population by age group}

\item{contact_mat}{a contact matrix from \code{\link[squire]{get_mixing_matrix}}}

\item{time_period}{maximum time of simulation}

\item{max_age}{max age for humans}

\item{lambda_external}{a vector of length equal to \code{time_period} giving an additional additive term contributing to the force of infection.
Models infectious contacts the population has with external (unmodeled) populations.}

\item{dt}{time step size}

\item{...}{Other parameters for \link[nimue:parameters]{nimue::parameters}}
}
\description{
This calls \code{\link[nimue]{parameters}} from \href{https://mrc-ide.github.io/nimue/index.html}{nimue}.
}
