% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_proportion_vaccinated_all_ages_cpp}
\alias{get_proportion_vaccinated_all_ages_cpp}
\title{Get proportion of an age group that is vaccinated (multi-dose, no types)}
\usage{
get_proportion_vaccinated_all_ages_cpp(variables, N_age, dose)
}
\arguments{
\item{variables}{a list}

\item{N_age}{number of age groups}

\item{dose}{which dose to get proportion of age group who has received it (NOT 0-indexed)}
}
\description{
Get proportion all age groups that have received a particular vaccine dose
by this timestep. This is similar to the function \code{\link[nimue]{coverage}} in the nimue package.
}
