% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_vaccination.R
\name{get_proportion_vaccinated_nimue}
\alias{get_proportion_vaccinated_nimue}
\title{Get proportion of an age group that is vaccinated (nimue vaccine model)}
\usage{
get_proportion_vaccinated_nimue(variables, age)
}
\arguments{
\item{variables}{a list}

\item{age}{an age group (can be a vector of multiple age groups)}
}
\description{
Get proportion of an age group that has received a particular vaccine dose
by this timestep.
}
