% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vaccination.R
\name{get_proportion_vaccinated_type}
\alias{get_proportion_vaccinated_type}
\title{Get proportion of an age group that is vaccinated (multi-dose, with types)}
\usage{
get_proportion_vaccinated_type(variables, age, type, dose)
}
\arguments{
\item{variables}{a list}

\item{age}{an age group (can be a vector of Multiple age groups)}

\item{type}{type of vaccine}

\item{dose}{which dose to get proportion of age group who has received it}
}
\description{
Get proportion of an age group that has received a particular vaccine type and dose
by this timestep. This is similar to the function \code{\link[nimue]{coverage}} in the nimue package.
}
