% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_states.R
\name{get_state_vector}
\alias{get_state_vector}
\title{Get vector of states from parameters}
\usage{
get_state_vector(psq)
}
\arguments{
\item{psq}{the model parameters}
}
\value{
a named vector
}
\description{
Get initial state conditions of the model from parameters.
}
