% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_nat.R
\name{get_time_since_last_dose_or_infection}
\alias{get_time_since_last_dose_or_infection}
\title{Calculate the time elapsed in days since each person's last dose or infection}
\usage{
get_time_since_last_dose_or_infection(
  timestep,
  dt,
  vaccinated_or_infected,
  dose_time,
  inf_time
)
}
\arguments{
\item{timestep}{current time step}

\item{dt}{size of time step}

\item{vaccinated_or_infected}{\link[individual:Bitset]{individual::Bitset} of vaccinated persons}

\item{dose_time}{an \link[individual:IntegerVariable]{individual::IntegerVariable} object}

\item{inf_time}{an \link[individual:IntegerVariable]{individual::IntegerVariable} object}
}
\description{
Calculate the time elapsed in days since each person's last dose or infection
}
