% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_nat.R
\name{get_time_since_last_infection}
\alias{get_time_since_last_infection}
\title{Calculate the time elapsed in days since each person's last infection}
\usage{
get_time_since_last_infection(timestep, dt, infected, inf_time)
}
\arguments{
\item{timestep}{current time step}

\item{dt}{size of time step}

\item{infected}{\link[individual:Bitset]{individual::Bitset} of infected persons}

\item{inf_time}{an \link[individual:IntegerVariable]{individual::IntegerVariable} object}
}
\description{
Calculate the time elapsed in days since each person's last infection
}
