% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines.R
\name{get_vaccination_priority_stage}
\alias{get_vaccination_priority_stage}
\title{Get prioritisation step for a specific dosing phase}
\usage{
get_vaccination_priority_stage(variables, phase, parameters)
}
\arguments{
\item{variables}{a list}

\item{phase}{current dosing phase}

\item{parameters}{a list}
}
\description{
Return which row of the \code{\link[nimue]{strategy_matrix}} the vaccination
program should be targeting for coverage. To complete a step, all \code{phase}
dose coverage should be >= prioritisation matrix target and all \code{phase + 1}
dose coverage for prioritized groups should be >= prioritisation matrix target.
If these two conditions are fulfilled for the entire \code{phase}, the function returns
-1 to indicate that vaccination dosing \code{phase} should be advanced.
}
