% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_vaccine.R
\name{get_vaccine_ab_titre_parameters}
\alias{get_vaccine_ab_titre_parameters}
\title{Get vaccine efficacy and Ab titre parameters}
\usage{
get_vaccine_ab_titre_parameters(max_dose = 2)
}
\description{
Get parameters for vaccine efficacy and antibody titre decay rate.
}
