% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_vector_cpp}
\alias{get_vector_cpp}
\title{Get a value from a vector}
\usage{
get_vector_cpp(vector_set, i)
}
\arguments{
\item{vector_set}{the set of values}

\item{i}{the day (assumes zero indexing)}
}
\description{
Get a value at some specific day
}
