% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{identify_ages_to_adjust}
\alias{identify_ages_to_adjust}
\title{Identify ages to adjust}
\usage{
identify_ages_to_adjust(discrete_ages, parameters)
}
\arguments{
\item{discrete_ages}{discrete ages}

\item{parameters}{Parameters object created by \code{\link{get_parameters}}}
}
\value{
Returns values to swap
}
\description{
Identify ages to adjust
}
\details{
Identifies age variables to be switched based on parameters object
}
