% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_nat.R
\name{independent_ab_titre_process}
\alias{independent_ab_titre_process}
\title{Update NAT for seperate vaccine and infection-derived NAT}
\usage{
independent_ab_titre_process(parameters, variables, dt)
}
\arguments{
\item{parameters}{a list of model parameters}

\item{variables}{a list of model variables}

\item{dt}{time step size}
}
\description{
Process that updates NAT titre for model where vaccine and infection derived
NAT are stored separately.
The NAT values are stored on the natural log scale.
This process will not calculate decay correctly for \code{dt > 1} so that is disallowed.
}
