% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_infection.R
\name{infection_process}
\alias{infection_process}
\title{Infection process (squire transmission model)}
\usage{
infection_process(parameters, variables, events, dt)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{a list of model variables, the output of \link{create_variables}}

\item{events}{a list of \link[individual:TargetedEvent]{individual::TargetedEvent}, the output of \link{create_events}}

\item{dt}{the time step}
}
\description{
simulates the process of infection.
}
