% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_infection.R
\name{infection_process_cpp}
\alias{infection_process_cpp}
\title{C++ infection process (squire model)}
\usage{
infection_process_cpp(parameters, variables, events, dt)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{dt}{the time step}
}
\description{
Simulates the infection process for the squire transmission model.
Calls \code{\link{infection_process_cpp_internal}} to return an external pointer object.
}
