% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infection_process_cpp_internal}
\alias{infection_process_cpp_internal}
\title{C++ infection process (squire model)}
\usage{
infection_process_cpp_internal(parameters, states, discrete_age, exposure, dt)
}
\description{
this is an internal function, you should use the R interface
for type checking, \code{\link{infection_process_cpp}}
}
