% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_infection_nimue.R
\name{infection_process_nimue}
\alias{infection_process_nimue}
\title{Transitions from S (nimue vaccine model)}
\usage{
infection_process_nimue(parameters, variables, events, dt)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{dt}{the time step}
}
\description{
S -> E if infected; can also become vaccinated. This incorporates
the slightly more complex force of infection calculation from the nimue model.
}
