% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_infection_nimue.R
\name{infection_process_nimue_cpp}
\alias{infection_process_nimue_cpp}
\title{C++ infection process (nimue vaccine model)}
\usage{
infection_process_nimue_cpp(parameters, variables, events, dt)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{dt}{the time step}
}
\description{
Simulates the infection process for the nimue vaccine model.
Calls \code{\link{infection_process_nimue_cpp_internal}} to return an external pointer object.
}
