% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infection_process_nimue_cpp_internal}
\alias{infection_process_nimue_cpp_internal}
\title{C++ infection process (nimue vaccine model)}
\usage{
infection_process_nimue_cpp_internal(
  parameters,
  states,
  vaccine_states,
  discrete_age,
  exposure,
  dt
)
}
\arguments{
\item{parameters}{a list of parameters from \code{\link{get_parameters_nimue}}}

\item{states}{a \code{\link[individual]{CategoricalVariable}}}

\item{vaccine_states}{a \code{\link[individual]{IntegerVariable}}}

\item{discrete_age}{a \code{\link[individual]{IntegerVariable}}}

\item{exposure}{a \code{\link[individual]{TargetedEvent}}}

\item{dt}{size of time step}
}
\description{
this is an internal function, you should use the R interface
for type checking, \code{\link{infection_process_cpp}}
}
