% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_infection_vaccine.R
\name{infection_process_vaccine}
\alias{infection_process_vaccine}
\title{Infection process for vaccine model (multi-dose, no types)}
\usage{
infection_process_vaccine(parameters, variables, events, dt)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{dt}{the time step}
}
\description{
This samples infection events in the susceptible population.
}
