% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_infection_variants.R
\name{infection_process_voc}
\alias{infection_process_voc}
\title{Infection process for vaccine model with variants of concern (multi-dose, no types)}
\usage{
infection_process_voc(parameters, variables, events, dt)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{dt}{the time step}
}
\description{
This samples infection events in the susceptible population.
}
