% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-infection-process.R
\name{infection_process_zzz}
\alias{infection_process_zzz}
\title{Transitions from S}
\usage{
infection_process_zzz(parameters, variables, events, dt, vaccines = NULL)
}
\arguments{
\item{variables}{Model variable}

\item{events}{Model events}

\item{paramaters}{Model parameters}
}
\description{
S -> E if infected; can also become vaccinated.
}
