% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_vaccines.R
\name{initialize_vaccine_variables}
\alias{initialize_vaccine_variables}
\title{Initialize vaccination variables (multi-dose, no types)}
\usage{
initialize_vaccine_variables(variables, dose_time_init, dose_num_init)
}
\arguments{
\item{variables}{a list from \code{\link{create_vaccine_variables}}}

\item{dose_time_init}{list of dose times (must all be integer vectors of the same length, -1 is value to indicate that individual has not
yet received this dose)}

\item{dose_num_init}{vector of initial doses (must be self-consistent with \code{dose_time_init})}
}
\description{
Initialize vaccination variables (multi-dose, no types)
}
