% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{integer_count_render_process}
\alias{integer_count_render_process}
\title{Render integer variables}
\usage{
integer_count_render_process(renderer, variable, margin)
}
\arguments{
\item{renderer}{a \code{\link[individual]{Render}} object}

\item{variable}{a \code{\link[individual]{IntegerVariable}} object}

\item{margin}{the set of values which may be taken for the integer variable}
}
\description{
Renders the number of individuals in each integer count. Should only be used
for integer variables with bounded ranges
}
