% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{integer_count_render_process_daily}
\alias{integer_count_render_process_daily}
\title{Render integer variables every day}
\usage{
integer_count_render_process_daily(renderer, variable, margin, dt)
}
\arguments{
\item{renderer}{a \code{\link[individual]{Render}} object}

\item{variable}{a \code{\link[individual]{IntegerVariable}} object}

\item{margin}{the set of values which may be taken for the integer variable}

\item{dt}{size of time step}
}
\description{
Renders the number of individuals in each integer count. Should only be used
for integer variables with bounded ranges. This only renders output on timesteps
that correspond to a day. This saves memory and potentially time for
simulations with small timesteps.
}
