% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{interp_input_par}
\alias{interp_input_par}
\title{Interpolate input parameters}
\usage{
interp_input_par(x, y, by = 1, end = max(x))
}
\arguments{
\item{x}{Time points for \link{y} changing}

\item{y}{Object to be interpolated}

\item{by}{Time steps to interpolate at. Default = 1}

\item{end}{End time point for interpoation. Default = \code{max(x)}}
}
\description{
Interpolate input parameters
}
\details{
Constant interpolation of time changing parameters
}
