% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_rt}
\alias{interpolate_rt}
\title{Produce piecewise linear Rt}
\usage{
interpolate_rt(dates, rt, max_date = NULL)
}
\arguments{
\item{dates}{a vector of \link{Date}s}

\item{rt}{a vector of reproductive values at those dates}

\item{max_date}{the maximum date of simulation (optional)}
}
\value{
a \link{list} with named elements \code{Rt} and \code{Rt_tt}, which can be provided
to the \link{get_parameters} function's arguments \code{R0} and \code{tt_R0}.
}
\description{
Make a piecewise linear interpolating vector of Rt, for each
day within the simulated interval. The assumed start of simulation
is the first date provided in \code{dates}, and the end of simulation is the last
date of \code{dates} unless \code{max_date} is provided. Interpolation is done with
inclusive endpoints.
}
