% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso3c_ages}
\alias{iso3c_ages}
\title{Median ages for countries}
\format{
A dataframe of 3 elements:

\code{$iso3c_ages}: A dataframe of iso3c countries and median ages
\itemize{
\item country: Country name string
\item age: Median population age numeric
\item iso3c: ISO3C string
}
}
\source{
\url{https://www.worldometers.info/world-population/population-by-country/}
}
\usage{
iso3c_ages
}
\description{
Median ages for countries
}
\keyword{datasets}
