% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_nat.R
\name{make_immune_parameters}
\alias{make_immune_parameters}
\title{Attach parameters dealing with immune (NAT) dynamics to parameter list}
\usage{
make_immune_parameters(
  parameters,
  vfr,
  mu_ab_infection = NULL,
  std10_infection = NULL
)
}
\arguments{
\item{parameters}{a list of model parameters}

\item{vfr}{a vector from \link{variant_fold_reduction_vector}}

\item{mu_ab_infection}{a vector of values associated with NAT for each infection bout;
please read \link{attach_event_listeners_natural_immunity} carefully depending on
if you are using the additive or overwriting NAT model, because this parameter's
interpretation will change. It may also be given as a vector, with each column
indexing a day in the simulation (so it is equal to \code{time_period}).}

\item{std10_infection}{a standard deviation for log-normal draw of NAT values}
}
\description{
Attach parameters dealing with immune (NAT) dynamics to parameter list
}
