% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_nat.R
\name{make_independent_vaccine_infection_nat_parameters}
\alias{make_independent_vaccine_infection_nat_parameters}
\title{Attach parameters for modeling seperate vaccine and infection derived NAT}
\usage{
make_independent_vaccine_infection_nat_parameters(
  parameters,
  dr_vec_doses = NULL,
  dr_vec_inf = NULL,
  max_ab_inf = NULL
)
}
\arguments{
\item{parameters}{a list of model parameters}

\item{dr_vec_doses}{a matrix where each column gives the decay rate vector of NATs following a
particular dose, and each row is a day}

\item{dr_vec_inf}{a vector where each element is the decay rate of NATs following infection}

\item{max_ab_inf}{maximum allowable NAT boost from natural infection}
}
\description{
Attach parameters for modeling seperate vaccine and infection derived NAT
}
