% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_rexp}
\alias{make_rexp}
\title{Make discretized exponential waiting time distribution}
\usage{
make_rexp(mu, dt, shift = 0L)
}
\arguments{
\item{mu}{Mean duration}

\item{dt}{Size of time step}

\item{shift}{number of time steps to add to sampled value}
}
\description{
Make geometric approximation to
continuous time exponential distribution.
}
