% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_vaccine.R
\name{make_vaccine_parameters}
\alias{make_vaccine_parameters}
\title{Combine and verify vaccine parameters}
\usage{
make_vaccine_parameters(
  safir_parameters,
  vaccine_ab_parameters,
  vaccine_set,
  dose_period,
  strategy_matrix,
  next_dose_priority_matrix
)
}
\arguments{
\item{safir_parameters}{a list from \code{\link{get_parameters}}}

\item{vaccine_ab_parameters}{a list from \code{\link{get_vaccine_ab_titre_parameters}}}

\item{vaccine_set}{a vector giving the number of doses available each day (not each timestep)}

\item{dose_period}{a vector giving the minimum delay between doses}

\item{strategy_matrix}{a vaccine strategy matrix from \code{\link[nimue]{strategy_matrix}}
or a list of strategy matrices, one for each vaccine dose}

\item{next_dose_priority_matrix}{a binary matrix giving age groups prioritized for next dose;
it should have one fewer row than the number of doses being given, because on the
final allocation phase there will be no future dose to prioritize}
}
\description{
Combine parameters for simulation and verify for correctness.
}
