% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_2vec_mult_cpp}
\alias{matrix_2vec_mult_cpp}
\title{Multiply a matrix by a integer vector and a double vector}
\usage{
matrix_2vec_mult_cpp(m, a, b)
}
\arguments{
\item{m}{a matrix}

\item{a}{a vector of int (must have length equal to number of columns of \code{m})}

\item{b}{a vector of double (must have length equal to number of columns of \code{m})}
}
\description{
Multiply a matrix by a integer vector and a double vector
}
