% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_vec_mult_cpp}
\alias{matrix_vec_mult_cpp}
\title{Multiply a matrix by a integer vector}
\usage{
matrix_vec_mult_cpp(m, a)
}
\arguments{
\item{m}{a matrix}

\item{a}{a vector (must have length equal to number of columns of \code{m})}
}
\description{
Multiply a matrix by a integer vector
}
