% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult_2matrix_rowsum}
\alias{mult_2matrix_rowsum}
\title{Element-wise multiply two matrices and take row sums}
\usage{
mult_2matrix_rowsum(a, b)
}
\arguments{
\item{a}{a matrix}

\item{b}{a matrix}
}
\description{
This function does no argument checking, please make sure \code{a} and \code{b}
are matrices of the same dimension.
}
