% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-hosp.R
\name{process_hosp_renderers}
\alias{process_hosp_renderers}
\title{Process rendering output of hospitalization and ICU incidence}
\usage{
process_hosp_renderers(renderers, parameters)
}
\arguments{
\item{renderers}{a named list of \link[individual:Render]{individual::Render} objects from \link{create_hosp_renderers}}

\item{parameters}{a named list of model parameters}
}
\description{
Process output from each rendering object and return a single
\link[data.table:data.table]{data.table::data.table} object, where incidence has been aggregated by
day.
}
