% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_erlang}
\alias{r_erlang}
\title{Erlang waiting time distribution}
\usage{
r_erlang(size, mu)
}
\arguments{
\item{size}{Number of draws}

\item{mu}{Mean duration}
}
\description{
Erlang waiting time distribution
}
\details{
Random draws from erlang distribution
}
