% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_exp}
\alias{r_exp}
\title{Exponential waiting time distribution}
\usage{
r_exp(size, mu)
}
\arguments{
\item{size}{Number of draws}

\item{mu}{Mean duration}
}
\description{
Exponential waiting time distribution
}
\details{
Random draws from exponential distribution
}
