% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(pop, parameters)
}
\arguments{
\item{pop}{population. See \link[squire:get_population]{squire::get_population}}

\item{parameters}{model parameters}
}
\description{
The main entrypoint for the simulation. run_simulation puts together the
model components and runs the malaria simulation. This currently returns a
dataframe with the number of individuals in each state at each timestep
}
