% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_simulation_replicate}
\alias{run_simulation_replicate}
\title{Run the simulation with repetitions}
\usage{
run_simulation_replicate(repetitions, overrides = list())
}
\arguments{
\item{repetitions}{n times to run the simulation}

\item{overrides}{a named list of parameters to use instead of defaults}
}
\value{
data frame for runs
}
\description{
Run the simulation with repetitions
}
