% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safir.R
\docType{package}
\name{safir}
\alias{safir}
\alias{safir-package}
\title{safir: squire and friends individual rewrite}
\description{
safir is an individual-based simulation model of COVD-19 based on squire.
}
\author{
\strong{Maintainer}: OJ Watson \email{o.watson15@imperial.ac.uk}

Other contributors:
\itemize{
  \item Imperial College of Science, Technology and Medicine [copyright holder]
}

}
