% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_vaccination.R
\name{schedule_dose_vaccine}
\alias{schedule_dose_vaccine}
\title{Schedule some individuals for a vaccination dose}
\usage{
schedule_dose_vaccine(
  timestep,
  variables,
  target,
  dose,
  parameters,
  correlated = FALSE
)
}
\arguments{
\item{timestep}{current time step}

\item{variables}{a list}

\item{target}{a \code{\link[individual]{Bitset}}}

\item{dose}{which dose}

\item{parameters}{model parameters}

\item{correlated}{are doses correlated?}
}
\description{
This is called from the event listeners for each dose, and also
aids in better testing of the simulation model
}
