% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_hospital.R
\name{schedule_outcome}
\alias{schedule_outcome}
\title{Schedule outcome}
\usage{
schedule_outcome(target, prob_successful, success_event, failure_event)
}
\arguments{
\item{target}{the individuals to draw from}

\item{prob_successful}{the probability each target individual is successful}

\item{success_event}{will be scheduled on success}

\item{failure_event}{will be scheduled on failure}
}
\description{
schedule individuals into follow up events based based on bernoulli draws of
\code{prob_successful}. It does not modify the input bitset \code{target}.
}
