% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_events.R
\name{setup_events}
\alias{setup_events}
\title{Schedule events for individuals at initialisation}
\usage{
setup_events(parameters, events, variables, dt)
}
\arguments{
\item{parameters}{model parameters}

\item{events}{a list of \link[individual:TargetedEvent]{individual::TargetedEvent}, the output of \link{create_events}}

\item{variables}{a list of model variables, the output of \link{create_variables}}

\item{dt}{size of the time step}
}
\description{
Schedule events for individuals at initialisation
}
