% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_events_nimue.R
\name{setup_events_nimue}
\alias{setup_events_nimue}
\title{Schedule events for individuals at initialisation (nimue vaccine model)}
\usage{
setup_events_nimue(parameters, events, variables, dt)
}
\arguments{
\item{parameters}{model parameters from \link{get_parameters_nimue}}

\item{events}{model events}

\item{variables}{model variables}

\item{dt}{size of time step}
}
\description{
Schedule events for individuals at initialisation (nimue vaccine model)
}
