% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_events_vaccine.R
\name{setup_events_vaccine}
\alias{setup_events_vaccine}
\title{Schedule events for individuals at initialisation (vaccine model, multi-dose, no types)}
\usage{
setup_events_vaccine(parameters, events, variables, dt)
}
\arguments{
\item{parameters}{write me!}

\item{events}{write me!}

\item{variables}{write me!}

\item{dt}{write me!}
}
\description{
Schedule events for individuals at initialisation (vaccine model, multi-dose, no types)
}
