% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_loop_safir.R
\name{simulation_loop_safir}
\alias{simulation_loop_safir}
\title{Simulation loop for safir models}
\usage{
simulation_loop_safir(
  variables = list(),
  events = list(),
  processes = list(),
  timesteps,
  variables_dont_update,
  progress = FALSE
)
}
\arguments{
\item{variables}{a list of Variables}

\item{events}{a list of Events}

\item{processes}{a list of processes to execute on each timestep}

\item{timesteps}{the number of timesteps to simulate}

\item{variables_dont_update}{character vector of variables that won't be updated}

\item{progress}{dispaly a progress bar?}
}
\description{
This is a replacement for \link[individual:simulation_loop]{individual::simulation_loop}
for models where not all values in \code{variables} need to be updated, and
also adds an optional progress bar
}
