% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_loop_vaccine.R
\name{simulation_loop_vaccine}
\alias{simulation_loop_vaccine}
\title{Simulation loop for safir (multi-dose, no types vaccine model)}
\usage{
simulation_loop_vaccine(
  variables = list(),
  events = list(),
  processes = list(),
  timesteps,
  progress = FALSE
)
}
\arguments{
\item{variables}{a list of Variables}

\item{events}{a list of Events}

\item{processes}{a list of processes to execute on each timestep}

\item{timesteps}{the number of timesteps to simulate}

\item{progress}{dispaly a progress bar?}
}
\description{
a modification of \code{\link[individual]{simulation_loop}}
to deal with the presence of special values in the \code{variables} list, and
also adds an optional progress bar
}
