% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tab_bins}
\alias{tab_bins}
\title{Tabulate a vector of observations}
\usage{
tab_bins(a, nbins)
}
\arguments{
\item{a}{a set of observations}

\item{nbins}{number of bins}
}
\description{
Tabulate a vector \code{a} whose values fall into a set of integers
of maximum value \code{nbins}. This function does no argument checking so please
ensure the maximum value of observations is not greater than \code{nbins}.
}
\examples{
nbin <- 10
a <- sample.int(n = nbin,size = 100,replace = T)
tabulate(bin = a,nbins = nbin)
tab_bins(a = a,nbins = nbin)
}
