% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tab_bins_weighted}
\alias{tab_bins_weighted}
\title{Tabulate a weighted vector of observations}
\usage{
tab_bins_weighted(a, wt, nbins)
}
\arguments{
\item{a}{a set of observations}

\item{wt}{a set of weights}

\item{nbins}{number of bins}
}
\description{
Similar to \link{tab_bins} but instead of each observation
being implicitly given weight 1, it now has weight given by \code{wt[i]}.
}
