% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines.R
\name{target_pop}
\alias{target_pop}
\title{Target persons in each age group to vaccinate (multi-dose, no types)}
\usage{
target_pop(
  phase,
  variables,
  parameters,
  t,
  dt,
  prioritisation,
  vaxx_priority = NULL
)
}
\arguments{
\item{phase}{which dose?}

\item{variables}{a list}

\item{parameters}{Model parameters}

\item{t}{current time step}

\item{dt}{size of time step}

\item{prioritisation}{row of the prioritisation matrix for the current step}

\item{vaxx_priority}{row of the vaxx_priority matrix for age groups that should get the next dose phase while still on
the current phase}
}
\value{
a named list with \code{n_to_cover}: number to be vaccinated in each age group,
\code{eligible_age_bsets}: \code{\link[individual]{Bitset}} of eligible group in each age bin (from \code{\link{age_group_eligible_for_dose_vaccine}})
\code{eligible_age_counts}: size of eligible group in each age bin (from \code{\link{age_group_eligible_for_dose_vaccine}})
}
\description{
Target persons in each age group to vaccinate (multi-dose, no types)
}
