% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_make_calculate_nat_cpp}
\alias{test_make_calculate_nat_cpp}
\title{used for testing only, calculate NAT}
\usage{
test_make_calculate_nat_cpp(variables, parameters, index, day)
}
\arguments{
\item{variables}{model variables}

\item{parameters}{model parameters}

\item{index}{a bitset}

\item{day}{current day}
}
\value{
a numeric vector
}
\description{
used for testing only, calculate NAT
}
