% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_vaccination.R
\name{update_vaccine_variables}
\alias{update_vaccine_variables}
\title{Update vaccine variables}
\usage{
update_vaccine_variables(variables)
}
\arguments{
\item{variables}{a list}
}
\description{
This should be called from the simulation loop. It does not
update disease state.
}
