% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_variants.R
\name{update_voc_variables}
\alias{update_voc_variables}
\title{Update variant of concern variables}
\usage{
update_voc_variables(variables)
}
\arguments{
\item{variables}{a list}
}
\description{
This should be called from the simulation loop. It does not
update disease state.
}
