% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_vaccination.R
\name{vaccination_process}
\alias{vaccination_process}
\title{Vaccination distribution process (multiple doses, no types)}
\usage{
vaccination_process(parameters, variables, events, dt)
}
\arguments{
\item{parameters}{a named list}

\item{variables}{a named list}

\item{events}{a named list}

\item{dt}{size of time step}
}
\description{
This function distributes vaccine doses each day for the vaccine
model with multiple doses but no types. It distributes vaccines according to the
following strategy.
\itemize{
\item start at \code{phase = 1}, meaning we begin by distributing the first dose
\item use \code{\link{get_vaccination_priority_stage}} to check what stage of the
prioritization matrix \code{vaccine_coverage_mat} from \code{\link[nimue]{strategy_matrix}}
we should be using to decide who to vaccinate.
\item we can advance one row of the prioritization matrix when coverage for this dose (phase) >=
prioritization matrix and coverage for next dose (phase) >= prioritization matrix groups prioritized
for the next dose by the matrix \code{next_dose_priority}.
}
}
