% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_vaccination_nimue.R
\name{vaccination_process_nimue}
\alias{vaccination_process_nimue}
\title{Vaccination process  (nimue vaccine model)}
\usage{
vaccination_process_nimue(parameters, variables, events, dt)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{dt}{the time step}
}
\description{
This samples vaccination events (if there are vaccines available that day)
for individuals in S, E, R states.
}
