% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_vaccination_nimue.R
\name{vaccination_process_nimue_cpp}
\alias{vaccination_process_nimue_cpp}
\title{C++ vaccination process (nimue vaccine model)}
\usage{
vaccination_process_nimue_cpp(parameters, variables, events, dt)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{dt}{the time step}
}
\description{
This samples vaccination events (if there are vaccines available that day)
for individuals in S, E, R states.
Calls \code{\link{vaccination_process_nimue_cpp_internal}} to return an external pointer object.
}
