% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_vaccination.R
\name{vaccine_ab_titre_process}
\alias{vaccine_ab_titre_process}
\title{Process that updates the antibody (Ab) titre each time step}
\usage{
vaccine_ab_titre_process(parameters, variables, events, dt)
}
\arguments{
\item{ab_titre}{a \code{\link[individual]{DoubleVariable}} giving the antibody titre for the population}

\item{dose_time}{a list of \code{\link[individual]{IntegerVariable}} objects giving the time of doses for the population}

\item{dose_num}{a \code{\link[individual]{IntegerVariable}} giving the dose number of the population}
}
\description{
Process that updates the antibody (Ab) titre each time step
}
