% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vaccine_efficacy_infection_cpp}
\alias{vaccine_efficacy_infection_cpp}
\title{Compute vaccine efficacy against infection from Ab titre}
\usage{
vaccine_efficacy_infection_cpp(ab_titre, parameters)
}
\arguments{
\item{ab_titre}{a vector of Ab titres}

\item{parameters}{model parameters}
}
\description{
Compute vaccine efficacy against infection from Ab titre
}
