% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_vaccination.R
\name{vaccine_efficacy_infection_voc}
\alias{vaccine_efficacy_infection_voc}
\title{Compute vaccine efficacy against infection from Ab titre for each variant of concern}
\usage{
vaccine_efficacy_infection_voc(ab_titre, parameters)
}
\arguments{
\item{ab_titre}{a vector of Ab titres}

\item{parameters}{model parameters.}
}
\value{
a numeric matrix, 0 is maximally proective, 1 is maximally unprotective.
The rows correspond to each variant and columns are each individual.
}
\description{
Compute vaccine efficacy against infection from Ab titre for each variant of concern
}
