% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_vaccination.R
\name{vaccine_efficacy_severe}
\alias{vaccine_efficacy_severe}
\title{Compute vaccine efficacy against severe disease from Ab titre}
\usage{
vaccine_efficacy_severe(ab_titre, who)
}
\arguments{
\item{ab_titre}{a \code{\link[individual]{DoubleVariable}} containing values of Ab titre}

\item{who}{a \code{\link[individual]{Bitset}} telling who we should calculate efficacy against severe infection for}
}
\description{
Compute vaccine efficacy against severe disease from Ab titre
}
