% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vaccine_efficacy_severe_voc_cpp}
\alias{vaccine_efficacy_severe_voc_cpp}
\title{Compute vaccine efficacy against severe disease from Ab titre for each variant of concern (C++)}
\usage{
vaccine_efficacy_severe_voc_cpp(ab_titre, ef_infection, parameters)
}
\arguments{
\item{ab_titre}{a vector of Ab titres}

\item{ef_infection}{a vector of efficacy against infection from \code{\link{vaccine_efficacy_infection}}}

\item{parameters}{model parameters}
}
\value{
a numeric vector, 0 is maximally proective, 1 is maximally unprotective
}
\description{
This needs the efficacy against infection because efficacy against severe disease,
conditional on breakthrough infection is what safir needs, which is computed as  1 - ((1 - efficacy_disease)/(1 - efficacy_infection)).
}
