% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_nat.R
\name{vaccine_efficacy_transmission}
\alias{vaccine_efficacy_transmission}
\title{Compute vaccine efficacy against onward transmission from Ab titre}
\usage{
vaccine_efficacy_transmission(nat, parameters, day)
}
\arguments{
\item{nat}{a vector of NAT values}

\item{parameters}{model parameters.}

\item{day}{the current day}
}
\value{
a numeric vector, 0 is maximally protective, 1 is maximally unprotective
}
\description{
Compute vaccine efficacy against onward transmission from Ab titre
}
