% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vfr.R
\name{variant_fold_reduction_vector}
\alias{variant_fold_reduction_vector}
\title{Make vector of variant fold reduction in natural antibody titre values}
\usage{
variant_fold_reduction_vector(
  parameters,
  variables,
  dt,
  vfr,
  vfr_time_1,
  vfr_time_2
)
}
\arguments{
\item{parameters}{a list of model parameters}

\item{variables}{a list of model variables}

\item{dt}{time step size}

\item{vfr}{a list of model parameters}

\item{vfr_time_1}{a list of model variables}

\item{vfr_time_2}{time step size}
}
\description{
Generate a vector for each time step where no VFR effect corresponds
to values in the vector of \code{1}, and the appropriate VFR adjustment otherwise.
}
