% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_nimue.R
\name{append_vaccine_nimue}
\alias{append_vaccine_nimue}
\title{Append vaccine parameters from nimue model}
\usage{
append_vaccine_nimue(parameters, ...)
}
\arguments{
\item{parameters}{list from \link{get_parameters}}

\item{...}{Other parameters for \code{nimue:::}\code{\link[nimue]{parameters}}}
}
\description{
This calls \code{\link[nimue]{parameters}} from \href{https://mrc-ide.github.io/nimue/index.html}{nimue}
and appends the following parameters to the list:
\itemize{
\item gamma_vaccine_delay: mean duration of period from vaccination to vaccine protection.
\item gamma_V: mean duration of vaccine-derived immunity (days)
\item rel_infectiousness_vaccinated: Relative infectiousness per age category of vaccinated individuals relative to unvaccinated individuals. Default = rep(1, 17), which is no impact of vaccination on onwards transmissions
\item rel_infectiousness: Relative infectiousness per age category relative to maximum infectiousness category. Default = rep(1, 17)
\item vaccine_efficacy_infection: Efficacy of vaccine against infection. This parameter must either be length 1 numeric (a single efficacy for all age groups) or length 17 numeric vector (an efficacy for each age group). An efficacy of 1 will reduce FOI by 100 percent, an efficacy of 0.2 will reduce FOI by 20 percent etc.
\item prob_hosp: Efficacy of vaccine against severe (requiring hospitilisation) disease (by age). This parameter must either be length 1 numeric (a single efficacy for all age groups) or length 17 numeric vector (an efficacy for each age group). An efficacy of 1 will reduce the probability of hospitalisation by 100 percent, an efficacy of 0.2 will reduce the probability of hospitalisation by 20 percent etc.
\item vaccine_coverage_mat: Vaccine coverage targets by age (columns) and priority (row)
\item N_prioritisation_steps: number of vaccine prioritization levels
\item current_prioritisation_step: current prioritization step we are on
\item vaccine_set: vaccines available each day of simulation
}
}
