% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines.R
\name{assign_doses}
\alias{assign_doses}
\title{Assign N doses to age groups based on weightings by number of people eligible to be vaccinated (multi-dose, no types)}
\usage{
assign_doses(t, dt, doses, n_to_cover, variables, events, phase, parameters)
}
\arguments{
\item{t}{current time step}

\item{dt}{size of time step}

\item{doses}{Total available doses}

\item{n_to_cover}{Number of people eligible to be vaccinated in each age group, from \code{\link{target_pop}}}

\item{variables}{a list}

\item{events}{a list}

\item{phase}{vaccination phase (which doses are we administering)}

\item{parameters}{a list}
}
\description{
Please make sure you are subtracting these doses from some daily total outside of this function.
It combines functionality of \code{\link[nimue]{assign_doses}} and \code{\link[nimue]{administer_first_dose}}/\code{\link[nimue]{administer_second_dose}}.
}
