% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines.R
\name{assign_doses}
\alias{assign_doses}
\title{Assign N doses to age groups based on weightings by number of people eligible to be vaccinated (multi-dose, no types)}
\usage{
assign_doses(
  t,
  dt,
  doses,
  n_to_cover,
  eligible_age_bset,
  eligible_age_counts,
  variables,
  events,
  phase,
  parameters
)
}
\arguments{
\item{t}{current time step}

\item{dt}{size of time step}

\item{doses}{Total available doses}

\item{n_to_cover}{Number of people eligible to be vaccinated in each age group, from \code{\link{target_pop_new}}}

\item{eligible_age_bset}{a list of \code{\link[individual]{Bitset}} from \code{\link{target_pop_new}}}

\item{eligible_age_counts}{size of eligible group in each age bin from \code{\link{target_pop_new}}}

\item{variables}{a list}

\item{events}{a list}

\item{phase}{vaccination phase (which doses are we administering)}

\item{parameters}{a list}
}
\value{
the number of remaining doses
}
\description{
Please make sure you are subtracting these doses from some daily total outside of this function.
It combines functionality of \code{\link[nimue]{assign_doses}} and \code{\link[nimue]{administer_first_dose}}/\code{\link[nimue]{administer_second_dose}}.
Be aware that it is possible for fewer doses to be assigned than supplied, if not enough doses are available. In general
\code{n_to_cover} is the ideal allocation, and it is constrained by the amount available, \code{doses}
}
