% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_naturalimmunity.R
\name{attach_event_listeners_natural_immunity}
\alias{attach_event_listeners_natural_immunity}
\title{Attach event listeners for natural immunity events}
\usage{
attach_event_listeners_natural_immunity(variables, events, parameters, dt)
}
\arguments{
\item{variables}{a named list of variables}

\item{events}{a named list of events}

\item{parameters}{the parameters}

\item{dt}{size of time step}
}
\description{
Attach event listeners for natural immunity events
}
