% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{categorical_count_renderer_process_daily}
\alias{categorical_count_renderer_process_daily}
\title{Render categories every day}
\usage{
categorical_count_renderer_process_daily(renderer, variable, categories, dt)
}
\arguments{
\item{renderer}{a \code{\link[individual]{Render}} object}

\item{variable}{a \code{\link[individual]{CategoricalVariable}} object}

\item{categories}{a character vector of categories to render}

\item{dt}{size of time step}
}
\value{
a function which can be passed as a process to \code{\link{simulation_loop}}
}
\description{
Renders the number of individuals in each category, but only
on timesteps that correspond to a day. This saves memory and potentially time
for simulations with small timesteps.
}
