% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{compartments_age_render_process_daily}
\alias{compartments_age_render_process_daily}
\title{Render categories by age every day}
\usage{
compartments_age_render_process_daily(
  renderer,
  age,
  compartments,
  parameters,
  dt
)
}
\arguments{
\item{renderer}{a \link[individual:Render]{individual::Render} object}

\item{age}{a \link[individual:IntegerVariable]{individual::IntegerVariable} object}

\item{compartments}{a \link[individual:CategoricalVariable]{individual::CategoricalVariable} object}

\item{parameters}{model parameters}

\item{dt}{size of time step}
}
\description{
This only renders output on timesteps that correspond to a day.
See examples for how to quickly summarize the output.
}
\examples{
\dontrun{
# if the renderer object is called dose_age_renderer
tmp <- as.data.table(dose_age_renderer$to_dataframe())
tmp <- melt(tmp,id.vars="timestep")
tmp1 <- tmp[, tstrsplit(variable, "_", keep = c(2, 4))]
tmp[ , variable := NULL]
tmp <- cbind(tmp, tmp1)
setnames(x = tmp,old = c("V1", "V2"),new = c("dose","age"))
dose_dt <- tmp[, .(value = sum(value)), by = .(dose, timestep)]
}
}
