% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_vaccination_nimue.R
\name{create_events_nimue}
\alias{create_events_nimue}
\title{Create events}
\usage{
create_events_nimue(events, parameters, vaccines = NULL)
}
\arguments{
\item{events}{a named list of individual::Event}

\item{parameters}{write me!}

\item{vaccines}{not currently used}
}
\description{
Create events
}
