% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_vaccination_nimue.R
\name{create_v3v4_to_v5_listener_nimue}
\alias{create_v3v4_to_v5_listener_nimue}
\title{Create listener for decay of vaccine protection (nimue vaccine model)}
\usage{
create_v3v4_to_v5_listener_nimue(variables, events, parameters)
}
\arguments{
\item{variables}{a named list of variables}

\item{events}{a named list of individual::Event}

\item{parameters}{the parameters}
}
\description{
Create listener for decay of vaccine protection (nimue vaccine model)
}
