% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_vaccines.R
\name{create_vaccine_variables}
\alias{create_vaccine_variables}
\title{Create vaccination variables (multi-dose, no types)}
\usage{
create_vaccine_variables(variables, parameters)
}
\arguments{
\item{variables}{a list}

\item{parameters}{list of model parameters}
}
\value{
named list of individual::Variable
}
\description{
Create all individual variables for humans
}
