% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_vaccination.R
\name{create_vaccine_variables}
\alias{create_vaccine_variables}
\title{Create vaccination variables}
\usage{
create_vaccine_variables(
  variables,
  pop,
  vaxx_types = c("AZ", "MD", "PF", "JJ", "novax"),
  max_dose = 2
)
}
\arguments{
\item{variables}{a list}

\item{pop}{population list}

\item{vaxx_types}{a character vector of possible vaccine types}

\item{max_dose}{maximum number of possible doses}
}
\value{
named list of individual::Variable
}
\description{
Create all individual variables for humans
}
