% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-variables_vaccination.R
\name{create_vaccine_variables_nimue}
\alias{create_vaccine_variables_nimue}
\title{Create vaccination variables for nimue style vaccinations}
\usage{
create_vaccine_variables_nimue(variables, pop)
}
\arguments{
\item{variables}{a list}

\item{pop}{population list}
}
\value{
named list of individual::Variable
}
\description{
Create all individual variables for humans
}
