% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{create_variables}
\alias{create_variables}
\title{Create age & state variables}
\usage{
create_variables(pop, parameters)
}
\arguments{
\item{pop}{population list}

\item{parameters}{model parameters}
}
\value{
named list of variable objects
}
\description{
Create all age & state variables for simulated population
}
