% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cross_tab_margins}
\alias{cross_tab_margins}
\title{Cross tabulate two vectors with given margins}
\usage{
cross_tab_margins(a, b, a_margin, b_margin)
}
\arguments{
\item{a}{one set of observations}

\item{b}{another set of observations}

\item{a_margin}{number of distinct values of a (rows)}

\item{b_margin}{number of distinct values of b (cols)}
}
\description{
this is a replacement for \code{\link[base]{table}} that allows empty
cells because the margins have been specified. The input vectors \code{a} and \code{b}
must be the same length, this function does no argument checking.
}
\examples{
a <- 1:5
b <- c(1,2,3,1,2)
cross_tab_margins(a,b,5,3)
table(a,b)
}
