% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vaccination.R
\name{eligible_for_dose_vaccine}
\alias{eligible_for_dose_vaccine}
\title{Identity those persons eligible for a dose}
\usage{
eligible_for_dose_vaccine(dose, dose_period, variables, t, dt)
}
\arguments{
\item{dose}{which dose? (must be greater than 1)}

\item{dose_period}{days between \code{dose} and the previous dose (please make sure \code{dose / dt} produces an integer number of timesteps)}

\item{variables}{a list}

\item{t}{the current time step}

\item{dt}{size of the time step}
}
\value{
an \code{\link[individual]{Bitset}}
}
\description{
Find those individuals who have had the dose preceding \code{dose},
have not yet received the next one, and are beyond the \code{dose_period}.
This is similar to the function \code{\link[nimue]{eligable_for_second}} in the nimue package.
This function should only be called if simulation time is greater than the \code{dose_period}
}
