% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_nimue.R
\name{get_parameters_nimue}
\alias{get_parameters_nimue}
\title{Get parameters for safir (nimue vaccine model)}
\usage{
get_parameters_nimue(population, contact_mat, time_period, max_age = 100, ...)
}
\arguments{
\item{population}{integer vector of population by age group}

\item{contact_mat}{a contact matrix from \code{\link[squire]{get_mixing_matrix}}}

\item{time_period}{maximum time of simulation}

\item{max_age}{max age for humans}

\item{...}{Other parameters for \code{\link[nimue]{parameters}}}
}
\description{
This calls \code{\link[nimue]{parameters}} from \href{https://mrc-ide.github.io/nimue/index.html}{nimue}.
}
