% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines.R
\name{get_proportion_vaccinated}
\alias{get_proportion_vaccinated}
\title{Get proportion of an age group that is vaccinated (multi-dose, no types)}
\usage{
get_proportion_vaccinated(variables, age, dose)
}
\arguments{
\item{variables}{a list}

\item{age}{an age group (can be a vector of Multiple age groups)}

\item{dose}{which dose to get proportion of age group who has received it}
}
\description{
Get proportion of an age group that has received a particular vaccine dose
by this timestep. This is similar to the function \code{\link[nimue]{coverage}} in the nimue package.
}
