% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_vaccination.R
\name{get_time_since_last_dose}
\alias{get_time_since_last_dose}
\title{Calculate the time elapsed in days since each person's last dose}
\usage{
get_time_since_last_dose(
  timestep,
  dt,
  vaccinated,
  dose_num,
  dose_time,
  N_phase
)
}
\description{
Calculate the time elapsed in days since each person's last dose
}
