% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_nat.R
\name{get_time_since_last_dose}
\alias{get_time_since_last_dose}
\title{Calculate the time elapsed in days since each person's last dose}
\usage{
get_time_since_last_dose(timestep, dt, vaccinated, dose_time)
}
\arguments{
\item{timestep}{current time step}

\item{dt}{size of time step}

\item{vaccinated}{bitset of vaccinated persons}

\item{dose_time}{a list of \code{\link[individual]{IntegerVariable}} objects}
}
\description{
Calculate the time elapsed in days since each person's last dose
}
