% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_variants.R
\name{get_voc_parameters}
\alias{get_voc_parameters}
\title{Get variant of concern parameters}
\usage{
get_voc_parameters(
  tmax,
  voc_types = c("alpha", "beta", "delta", "wt", "none"),
  voc_trajectory = NULL
)
}
\arguments{
\item{tmax}{the maximum day of simulation (not the number of time steps)}

\item{voc_types}{character vector of variant names (must include "wt" and "none")}
}
\description{
Get variant of concern parameters
}
