% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_variants.R
\name{get_voc_parameters}
\alias{get_voc_parameters}
\title{Get variant of concern transmission parameters}
\usage{
get_voc_parameters(
  safir_parameters,
  voc_types = c("alpha", "beta", "delta", "wt", "none"),
  voc_trajectory = NULL,
  vaccine_parameters
)
}
\arguments{
\item{safir_parameters}{either a single object from \code{\link{get_parameters}}
or a list of objects from \code{\link{get_parameters}}}

\item{voc_types}{character vector of variant names (must include "wt" and "none")}

\item{vaccine_parameters}{either a single object from \code{\link{get_vaccine_ab_titre_parameters}}
or a list of objects from \code{\link{get_vaccine_ab_titre_parameters}} giving efficacy parameters
for each variant}
}
\description{
Get variant of concern transmission parameters
}
