% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-infection.R
\name{infection_process}
\alias{infection_process}
\title{Transitions from S}
\usage{
infection_process(parameters, variables, events, dt, vaccines = NULL)
}
\arguments{
\item{parameters}{Model parameters}

\item{variables}{Model variable}

\item{events}{Model events}

\item{dt}{the time step}

\item{vaccines}{vaccine parameters}
}
\description{
S -> E if infected; can also become vaccinated.
}
