% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_nat.R
\name{make_calculate_nat}
\alias{make_calculate_nat}
\title{Make function to calculate population NAT}
\usage{
make_calculate_nat(variables, parameters)
}
\arguments{
\item{variables}{a named list}

\item{parameters}{a named list}
}
\description{
The returned function takes arguments \code{index} (a bitset) and \code{day} and returns the NAT of persons in \code{index} on the linear scale.
}
