% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_rexp}
\alias{make_rexp}
\title{Make discretized exponential waiting time distribution}
\usage{
make_rexp(mu, dt, shift = 0L)
}
\arguments{
\item{mu}{Mean duration}

\item{dt}{Size of time step}

\item{shift}{number of time steps to add to sampled value}
}
\description{
Random draws from a geometric distribution
which approximates a continuous time exponential distribution.
See \href{https://kingaa.github.io/pomp/vignettes/C_API.html#exponentialgeometric-rate-conversion}{pomp documentation}
for details on the approximation.
}
