% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_rexp_simple}
\alias{make_rexp_simple}
\title{Make discretized exponential waiting time distribution (simple)}
\usage{
make_rexp_simple(mu, dt, shift = 0L)
}
\arguments{
\item{mu}{Mean duration}

\item{dt}{Size of time step}

\item{shift}{number of time steps to add to sampled value}
}
\description{
Make simple geometric approximation to
continuous time exponential distribution.
}
