% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_naturalimmunity.R
\name{natural_immunity_ab_titre_process}
\alias{natural_immunity_ab_titre_process}
\title{Process that updates the antibody (Ab) titre each time step for model vaccine model with natural immunity}
\usage{
natural_immunity_ab_titre_process(parameters, variables, dt)
}
\arguments{
\item{parameters}{a list of model parameters}

\item{variables}{a list of model variables}

\item{dt}{time step size}
}
\description{
The values in \code{ab_titre} are calculated on the log scale.
This process will not calculate decay correctly for \code{dt > 1} so that is disallowed.
}
