% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safir.R
\docType{package}
\name{safir}
\alias{safir}
\alias{safir-package}
\title{safir: squire and friends individual rewrite}
\description{
safir is an individual-based simulation model of COVD-19 based on squire.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrc-ide/safir}
  \item \url{https://mrc-ide.github.io/safir/}
  \item Report bugs at \url{https://github.com/mrc-ide/safir/issues}
}

}
\author{
\strong{Maintainer}: OJ Watson \email{o.watson15@imperial.ac.uk}

Authors:
\itemize{
  \item Peter Winskill \email{p.winskill@imperial.ac.uk}
  \item Sean L. Wu \email{slwood89@gmail.com} (\href{https://orcid.org/0000-0002-5781-9493}{ORCID})
}

Other contributors:
\itemize{
  \item Imperial College of Science, Technology and Medicine [copyright holder]
}

}
