% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-initialize.R
\name{setup_events}
\alias{setup_events}
\title{Schedule events for individuals at initialisation}
\usage{
setup_events(parameters, events, variables, dt, vaccines = NULL)
}
\description{
Schedule events for individuals at initialisation
}
