% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{setup_events}
\alias{setup_events}
\title{Schedule events for individuals at initialisation}
\usage{
setup_events(parameters, events, variables, dt, vaccines = NULL)
}
\arguments{
\item{parameters}{write me!}

\item{events}{write me!}

\item{variables}{write me!}

\item{dt}{write me!}

\item{vaccines}{write me!}
}
\description{
Schedule events for individuals at initialisation
}
