% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_events_nimue.R
\name{setup_events_nimue}
\alias{setup_events_nimue}
\title{Schedule events for individuals at initialisation (nimue vaccine model)}
\usage{
setup_events_nimue(parameters, events, variables, dt)
}
\arguments{
\item{parameters}{write me!}

\item{events}{write me!}

\item{variables}{write me!}

\item{dt}{write me!}
}
\description{
Schedule events for individuals at initialisation (nimue vaccine model)
}
