% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_events_vaccine.R
\name{setup_events_vaccine}
\alias{setup_events_vaccine}
\title{Schedule events for individuals at initialisation (vaccine model, multi-dose, no types)}
\usage{
setup_events_vaccine(parameters, events, variables, dt)
}
\arguments{
\item{parameters}{model parameters}

\item{events}{list of model events}

\item{variables}{list of model variables}

\item{dt}{size of time step}
}
\description{
Schedule events for individuals at initialisation (vaccine model, multi-dose, no types)
}
