% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines_fix.R
\name{target_pop}
\alias{target_pop}
\title{Target persons in each age group to vaccinate (multi-dose, no types)}
\usage{
target_pop(
  dose,
  variables,
  events,
  parameters,
  timestep,
  dt,
  strategy_matrix_step,
  next_dose_priority = NULL
)
}
\arguments{
\item{dose}{what dose to target (the phase we are vaccinating)}

\item{variables}{a list of model variables}

\item{events}{a list of model events}

\item{parameters}{a list of parameters}

\item{timestep}{the current time step}

\item{dt}{the size of the time step}

\item{strategy_matrix_step}{the current step of the strategy matrix}

\item{next_dose_priority}{(optional) a row of the next dose priority matrix}
}
\description{
Given the current dosing phase (\code{dose}), the current step of the strategy matrix (\code{strategy_matrix_step})
and possible the age groups prioritized for the next dose \code{next_dose_priority}, figure out who should get this \code{dose}.
This will return a list of bitsets giving people who are eligible to get this dose, taking into account that they
have not yet had it, nor are scheduled for it, and (if dose > 1) are past the threshold for their previous dose.
}
