% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines.R
\name{target_pop}
\alias{target_pop}
\title{Target number of individual in each age group to vaccinate (multi-dose, no types)}
\usage{
target_pop(
  phase,
  variables,
  parameters,
  t,
  dt,
  prioritisation,
  vaxx_priority = NULL
)
}
\arguments{
\item{phase}{which dose?}

\item{variables}{a list}

\item{parameters}{Model parameters}

\item{t}{current time step}

\item{dt}{size of time step}

\item{prioritisation}{row of the prioritisation matrix for the current step}

\item{vaxx_priority}{row of the vaxx_priority matrix for age groups that should get the next dose phase while still on
the current phase}
}
\description{
Target number of individual in each age group to vaccinate (multi-dose, no types)
}
