% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_vaccines.R
\name{target_pop}
\alias{target_pop}
\title{Target persons in each age group to vaccinate (multi-dose, no types)}
\usage{
target_pop(
  dose,
  variables,
  parameters,
  t,
  dt,
  prioritisation,
  vaxx_priority = NULL
)
}
\arguments{
\item{dose}{which dose?}

\item{variables}{a list}

\item{parameters}{Model parameters}

\item{t}{current time step}

\item{dt}{size of time step}

\item{prioritisation}{row of the prioritisation matrix for the current step}

\item{vaxx_priority}{vaxx_priority matrix for age groups that should get the next dose
the current dose}
}
\value{
a named list with \code{n_to_cover}: number to be vaccinated in each age group,
\code{eligible_age_bsets}: \code{\link[individual]{Bitset}} of eligible group in each age bin (from \code{\link{age_group_eligible_for_dose_vaccine}})
\code{eligible_age_counts}: size of eligible group in each age bin (from \code{\link{age_group_eligible_for_dose_vaccine}})
}
\description{
For a given dose, find eligible persons in each age group to target for that vaccine dose.
If \code{vaxx_priority} is specified, it means that we are currently vaccinating one dose ahead
of the current vaccination phase, for those age groups prioritized to get their next dose.
}
