% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vaccination_process_nimue_cpp_internal}
\alias{vaccination_process_nimue_cpp_internal}
\title{C++ infection process (nimue vaccine model)}
\usage{
vaccination_process_nimue_cpp_internal(
  parameters,
  states,
  eligible,
  vaccinated,
  empty,
  discrete_age,
  v0_to_v1v2,
  dt
)
}
\arguments{
\item{parameters}{a list of parameters from \code{\link{get_parameters_nimue}}}

\item{states}{a \code{\link[individual]{CategoricalVariable}}}

\item{eligible}{a \code{\link[individual]{Bitset}}}

\item{vaccinated}{a \code{\link[individual]{Bitset}}}

\item{empty}{a \code{\link[individual]{Bitset}}}

\item{discrete_age}{a \code{\link[individual]{IntegerVariable}}}

\item{v0_to_v1v2}{a \code{\link[individual]{TargetedEvent}}}

\item{dt}{size of time step}
}
\description{
this is an internal function, you should use the R interface
for type checking, \code{\link{infection_process_cpp}}
}
