% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_nat.R
\name{vaccine_efficacy_infection}
\alias{vaccine_efficacy_infection}
\title{Compute vaccine efficacy against infection from Ab titre}
\usage{
vaccine_efficacy_infection(ab_titre, parameters, day)
}
\arguments{
\item{ab_titre}{a vector of Ab titres}

\item{parameters}{model parameters.}

\item{day}{the current day}
}
\value{
a numeric vector, 0 is maximally proective, 1 is maximally unprotective
}
\description{
Compute vaccine efficacy against infection from Ab titre
}
