% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vaccine_efficacy_infection_cpp}
\alias{vaccine_efficacy_infection_cpp}
\title{Compute vaccine efficacy against infection from Ab titre (C++)}
\usage{
vaccine_efficacy_infection_cpp(nat, parameters, day)
}
\arguments{
\item{nat}{a vector of NAT values}

\item{parameters}{model parameters}

\item{day}{current day}
}
\value{
a numeric vector, 0 is maximally proective, 1 is maximally unprotective
}
\description{
Compute vaccine efficacy against infection from Ab titre (C++)
}
