% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_nat.R
\name{vaccine_efficacy_severe}
\alias{vaccine_efficacy_severe}
\title{Compute vaccine efficacy against severe disease from Ab titre}
\usage{
vaccine_efficacy_severe(nat, ef_infection, parameters, day)
}
\arguments{
\item{nat}{a vector of NAT values}

\item{ef_infection}{a vector of efficacy against infection from \code{\link{vaccine_efficacy_infection}}}

\item{parameters}{model parameters}

\item{day}{the current day}
}
\value{
a numeric vector, 0 is maximally proective, 1 is maximally unprotective
}
\description{
This needs the efficacy against infection because efficacy against severe disease,
conditional on breakthrough infection is what safir needs, which is computed as  1 - ((1 - efficacy_disease)/(1 - efficacy_infection)).
}
