% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vaccine_efficacy_transmission_cpp}
\alias{vaccine_efficacy_transmission_cpp}
\title{Compute vaccine efficacy against onward transmission from Ab titre (C++)}
\usage{
vaccine_efficacy_transmission_cpp(ab_titre, parameters, timestep)
}
\arguments{
\item{ab_titre}{a vector of Ab titres}

\item{parameters}{model parameters.}

\item{timestep}{current timestep}
}
\value{
a numeric vector, 0 is maximally protective, 1 is maximally unprotective
}
\description{
Compute vaccine efficacy against onward transmission from Ab titre (C++)
}
