% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_nat.R
\name{variant_fold_reduction_vector}
\alias{variant_fold_reduction_vector}
\title{Make vector of variant fold reduction in natural antibody titre values}
\usage{
variant_fold_reduction_vector(parameters, dt, vfr, vfr_time_1, vfr_time_2)
}
\arguments{
\item{parameters}{a list of model parameters}

\item{dt}{time step size}

\item{vfr}{a positive value giving the fold reduction}

\item{vfr_time_1}{time to begin reduction}

\item{vfr_time_2}{time to end reduction}
}
\description{
Generate a vector for each time step where no VFR effect corresponds
to values in the vector of \code{1}, and the appropriate VFR adjustment otherwise.
}
